+++ src/App.tsx
@@ -4,7 +4,7 @@ import Counter from './components/Counter';
 import { Button } from './components/ui/button';
 import { Signup } from './auth/sign-up';
 import Aside from './components/aside/aside';
-import Notice from './page/notice';
+import ShopArtistNotice from './page/shop-artist-notice';
 
 function MainPage() {
   const navigate = useNavigate();
@@ -36,7 +36,7 @@ function App() {
     <Routes>
       <Route path="/" element={<MainPage />} />
       <Route path="/signup" element={<Signup />} />
-      <Route path="/notice" element={<Notice />} />
+      <Route path="/notice" element={<ShopArtistNotice />} />
     </Routes>
   );
 }
